# brain-battle-hub/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brain Battle Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubmitScoreBody {
  username: string;
  gameId: string;
  score: number;
}

export interface Score {
  id: number;
  username: string;
  gameId: string;
  score: number;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  gameId: string;
  score: number;
  createdAt: string;
}

export interface LeaderboardStats {
  totalPlayers: number;
  topScore: number;
  gamesToday: number;
  topGame: string;
}

export type PlayerRankBadge =
  (typeof PlayerRankBadge)[keyof typeof PlayerRankBadge];

export const PlayerRankBadge = {
  bronze: "bronze",
  silver: "silver",
  gold: "gold",
} as const;

export interface PlayerRank {
  username: string;
  rank: number;
  totalPlayers: number;
  percentile: number;
  badge: PlayerRankBadge;
  bestScore: number;
  gamesPlayed: number;
}

export type GetLeaderboardParams = {
  gameId?: string;
  period?: GetLeaderboardPeriod;
  limit?: number;
};

export type GetLeaderboardPeriod =
  (typeof GetLeaderboardPeriod)[keyof typeof GetLeaderboardPeriod];

export const GetLeaderboardPeriod = {
  global: "global",
  daily: "daily",
} as const;

export type GetPlayerRankParams = {
  username: string;
};