# brain-battle-hub/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brain Battle Hub API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit a score after a game
 */
export const SubmitScoreBody = zod.object({
  username: zod.string(),
  gameId: zod.string(),
  score: zod.number(),
});

/**
 * @summary Get the global leaderboard
 */
export const GetLeaderboardQueryParams = zod.object({
  gameId: zod.coerce.string().optional(),
  period: zod.enum(["global", "daily"]).optional(),
  limit: zod.coerce.number().optional(),
});

export const GetLeaderboardResponseItem = zod.object({
  rank: zod.number(),
  username: zod.string(),
  gameId: zod.string(),
  score: zod.number(),
  createdAt: zod.string(),
});
export const GetLeaderboardResponse = zod.array(GetLeaderboardResponseItem);

/**
 * @summary Get top-level stats (total players, top score, games played today)
 */
export const GetLeaderboardStatsResponse = zod.object({
  totalPlayers: zod.number(),
  topScore: zod.number(),
  gamesToday: zod.number(),
  topGame: zod.string(),
});

/**
 * @summary Get rank info for a player by username
 */
export const GetPlayerRankQueryParams = zod.object({
  username: zod.coerce.string(),
});

export const GetPlayerRankResponse = zod.object({
  username: zod.string(),
  rank: zod.number(),
  totalPlayers: zod.number(),
  percentile: zod.number(),
  badge: zod.enum(["bronze", "silver", "gold"]),
  bestScore: zod.number(),
  gamesPlayed: zod.number(),
});