# brain-battle-hub/lib/api-zod/src/generated/types/playerRank.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brain Battle Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerRankBadge } from "./playerRankBadge";

export interface PlayerRank {
  username: string;
  rank: number;
  totalPlayers: number;
  percentile: number;
  badge: PlayerRankBadge;
  bestScore: number;
  gamesPlayed: number;
}