# brain-battle-hub/lib/api-zod/src/generated/types/submitScoreBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brain Battle Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitScoreBody {
  username: string;
  gameId: string;
  score: number;
}